% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Abilene_tbl_df}
\alias{Abilene_tbl_df}
\title{Crime Records of Abilene, Texas, USA}
\format{
A tibble with 16 observations and 3 variables:
\describe{
  \item{crimetype}{Type of crime (character).}
  \item{year}{Year of the reported crime (factor).}
  \item{number}{Number of reported crimes (integer).}
}
}
\source{
Uniform Crime Reports, U.S. Department of Justice.
}
\usage{
data(Abilene_tbl_df)
}
\description{
This dataset contains information on reported crimes in Abilene, Texas, including the type of crime,
year of the incident, and the number of reported cases. It provides a snapshot of crime patterns
in the city for the years 1992 and 1999.
}
\details{
The dataset name has been changed to 'Abilene_tbl_df' to avoid confusion with other data sets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
indicates that the dataset is a tibble in R. The original content has not been modified in any way.
}
