% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{camden_crimes_df}
\alias{camden_crimes_df}
\title{Crime Records of Camden Borough, UK}
\format{
A data frame with 4,578 observations and 4 variables:
\describe{
  \item{x}{X-coordinate (numeric).}
  \item{y}{Y-coordinate (numeric).}
  \item{date}{Date of the reported crime (Date).}
  \item{type}{Type of crime (character).}
}
}
\source{
Data comprising 'Theft' and 'Criminal Damage' records of Camden Borough of London, UK, 2021.
  (Source: \url{https://data.police.uk/data/})
}
\usage{
data(camden_crimes_df)
}
\description{
This dataset contains information on reported crimes in Camden, including spatial coordinates,
dates of the incidents, and crime types. It provides a detailed view of crime patterns
within the region.
}
\details{
The dataset name has been changed to 'camden_crimes_df' to avoid confusion with other data sets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'df'
indicates that the dataset is a traditional data frame in R. The original content has not been modified in any way
}
