% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimedatasets-package.R
\docType{package}
\name{crimedatasets}
\alias{crimedatasets}
\alias{crimedatasets-package}
\title{crimedatasets: A Comprehensive Collection of Crime-Related Datasets}
\description{
{
A comprehensive collection of datasets exclusively focused on crimes, criminal activities, and related topics. This package serves as a valuable resource for researchers, analysts, and students interested in crime analysis, criminology, and socio-economic studies related to criminal behavior.
}
}
\details{
crimedatasets: A Comprehensive Collection of Crime-Related Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Crime-Related Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/crimedatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Cáceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
