% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{crimtab_table}
\alias{crimtab_table}
\title{Student's 3000 Criminals Data}
\format{
A table with 42 rows and 22 columns:
\describe{
  \item{Var1}{Factor or categorical variable representing different crime categories.}
  \item{Var2}{A second factor or categorical variable, potentially representing different
  classifications such as location, time, or crime severity.}
  \item{Freq}{Frequency of occurrences within each combination of categories, representing
  the number of reported incidents for each combination.}
}
}
\source{
Public crime data.
}
\usage{
data(crimtab_table)
}
\description{
Data of 3000 male criminals over 20 years old undergoing their sentences
in the chief prisons of England and Wales.
}
\details{
The dataset name has been changed to 'crimtab_table' to avoid confusion with other data sets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics.
The suffix 'table' indicates that the dataset is stored as a contingency table, rather than
a traditional data frame. The original content has not been modified in any way.
}
