% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Gallup_tbl_df}
\alias{Gallup_tbl_df}
\title{Gallup Marijuana Possession Poll (1980)}
\format{
A tibble with 1,200 observations and 2 variables:
\describe{
  \item{demographics}{Demographic category of the respondent (factor with 12 levels).}
  \item{opinion}{Respondent's opinion on marijuana possession as a criminal offense (factor with 3 levels).}
}
}
\source{
Results of a Gallup poll conducted in 1980.
}
\usage{
data(Gallup_tbl_df)
}
\description{
This dataset contains the results of a Gallup poll conducted in 1980 regarding public opinion
on whether possession of marijuana should be considered a criminal offense. The dataset
includes demographic information and the corresponding opinions of the respondents.
}
\details{
The dataset name has been changed to 'Gallup_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets  package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
indicates that the dataset is stored as a tibble in R. The original content has not been modified in any way.
}
