% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Inmate_tbl_df}
\alias{Inmate_tbl_df}
\title{Type of Drug Offense by Race}
\format{
A tibble with 28,047 observations and 2 variables:
\describe{
  \item{race}{Race of the individual (factor with 3 levels).}
  \item{drug}{Type of drug offense (factor with 4 levels).}
}
}
\source{
C. Wolf Harlow (1994), *Comparing Federal and State Prison Inmates*, NCJ-145864, U.S. Department of Justice, Bureau of Justice Statistics.
}
\usage{
data(Inmate_tbl_df)
}
\description{
This dataset provides information on the type of drug offenses categorized by race. It contains
records that can be used to analyze racial patterns in drug-related offenses. The data is sourced
from a comparative study of federal and state prison inmates.
}
\details{
The dataset name has been changed to 'Inmate_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
indicates that the dataset is stored as a tibble. The original content has not been modified in any way.

This dataset provides insights into racial disparities and trends in drug offenses.
}
