% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Sentence_tbl_df}
\alias{Sentence_tbl_df}
\title{Sentences of 41 Prisoners Convicted of a Homicide Offense}
\format{
A tibble with 41 observations and 1 variable:
\describe{
  \item{months}{The number of months served in prison by each prisoner (integer).}
}
}
\source{
U.S. Department of Justice, Bureau of Justice Statistics, Prison Sentences and Time Served for Violence, NCJ-153858, April 1995.
}
\usage{
data(Sentence_tbl_df)
}
\description{
This dataset contains information on the length of sentences served by 41 prisoners convicted of a homicide offense.
The data was taken from a report by the U.S. Department of Justice, Bureau of Justice Statistics, which provides
insight into the sentencing and time served for violent crimes, specifically homicides.
The dataset includes the number of months each prisoner served in prison.
}
\details{
The dataset name has been changed to 'Sentence_tbl_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the crimedatasets package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' indicates that the dataset is a tibble.
The original content has not been modified in any way.
}
