% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{TerrorismGlobal_table}
\alias{TerrorismGlobal_table}
\title{Global Terrorism Database (GTD) Yearly Summaries}
\format{
A table with 10,200 rows and 50 columns:
\describe{
  \item{country_txt}{Character vector representing the country where the terrorist incident occurred.}
  \item{iyear}{Character vector representing the year the incident took place.}
}
}
\source{
Global Terrorism Database (GTD), 1970-2020.
}
\usage{
data(TerrorismGlobal_table)
}
\description{
This dataset contains yearly summaries of global terrorism incidents from 1970 onward.
The data includes information on over 209,000 incidents of terrorism, with details on the
country, year, and other relevant variables related to each incident.
}
\details{
The dataset name has been changed to 'TerrorismGlobal_table' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'table'
indicates that the dataset is represented as a table in R. The original content has not been modified in any way.
}
