% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USArrests_df}
\alias{USArrests_df}
\title{Violent Crime Rates by US State}
\format{
A data frame with 50 observations and 4 variables:
\describe{
  \item{Murder}{Murder arrests per 100,000 residents (numeric).}
  \item{Assault}{Assault arrests per 100,000 residents (integer).}
  \item{UrbanPop}{Percentage of the population living in urban areas (integer).}
  \item{Rape}{Rape arrests per 100,000 residents (numeric).}
}
}
\source{
1973 crime data, originally included in the USArrests dataset from R.
}
\usage{
data(USArrests_df)
}
\description{
This dataset contains statistics on violent crime rates in each of the 50 US states for the year 1973.
The data includes arrests per 100,000 residents for assault, murder, and rape, as well as the percentage of the population
living in urban areas.
}
\details{
The dataset name has been changed to 'USArrests_df' to maintain consistency with the naming conventions of the
crimedatasets package. The suffix 'df' indicates that the dataset is stored as a data frame in R.
The original content has not been modified in any way.
}
