% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USincarcerations_df}
\alias{USincarcerations_df}
\title{US Incarcerations 1925 Onward}
\format{
A data frame with 95 rows and 7 variables:
\describe{
  \item{year}{Year of the recorded data (numeric).}
  \item{stateFedIncarcerees}{Number of prisoners under state and federal jurisdiction (numeric).}
  \item{stateFedIncarcerationRate}{Incarceration rate per 100,000 population for state and federal facilities (numeric).}
  \item{stateFedMales}{Number of male prisoners in state and federal facilities (numeric).}
  \item{stateFedMaleRate}{Male incarceration rate per 100,000 male population (numeric).}
  \item{stateFedFemales}{Number of female prisoners in state and federal facilities (numeric).}
  \item{stateFedFemaleRate}{Female incarceration rate per 100,000 female population (numeric).}
}
}
\source{
US incarceration data (1925 onward).
}
\usage{
data(USincarcerations_df)
}
\description{
This dataset contains counts of prisoners under the jurisdiction of state and federal correctional authorities in the United States from 1925 onward.
The data excludes jail inmates and focuses on individuals in state and federal incarceration facilities.
}
\details{
The dataset name has been changed to 'USincarcerations_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'df'
indicates that the dataset is a data frame. The original content has not been modified in any way.
}
