% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{wmurders_ts}
\alias{wmurders_ts}
\title{Annual Female Murder Rate in the USA (1950-2004)}
\format{
A time series object with 55 observations and 1 variable:
\describe{
  \item{wmurders_ts}{Numeric vector representing the annual female murder rate per 100,000 population in the USA.}
}
}
\source{
U.S. crime statistics and historical records.
}
\usage{
data(wmurders_ts)
}
\description{
This dataset contains the annual female murder rate per 100,000 standard population in the United States from 1950 to 2004.
The data represents the total number of murdered women per 100,000 population on an annual basis, providing insights into
trends and patterns in female homicides over a period of 55 years.
}
\details{
The dataset name has been changed to 'wmurders_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
crimedatasets package and assists users in identifying its specific characteristics. The suffix 'ts'
indicates that the dataset is a time series object in R. The original content has not been modified in any way.
}
